refactor: scheduled task service
